% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{binmonodf}
\alias{binmonodf}
\title{Monotonic Binning Sample Data}
\format{Data frame with 5,000 rows and 6 columns.}
\description{
A simulated dataset used to illustrate the application of monotonic binning.
}
\details{
\itemize{
  \item fgood1: Default (0), Not Default (1) for Numeric Variable 1.
  \item chr1: Numeric variable 1.
  \item fgood2: Default (0), Not Default (1) for Numeric Variable 2.
  \item chr2: Numeric variable 2.
  \item fgood3: Default (0), Not Default (1) for Numeric Variable 3.
  \item chr3: Numeric variable 3.
  }
}
