% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{binTableCategorical}
\alias{binTableCategorical}
\title{Weight of Evidence and Information Value for a categorical variable}
\usage{
binTableCategorical(df, x, y, maxCategories = 5, verbose = FALSE)
}
\arguments{
\item{df}{Dataframe containing the two columns of data}

\item{x}{Name of the categorical variable as a string}

\item{y}{Name of the boolean response as a string}

\item{maxCategories}{maximum number of accepted categories. Default is 5.}

\item{verbose}{Boolean to add additional information. Default is \code{FALSE}}
}
\value{
A list with the total Information Value, and a tibble containing the bins with detailed
 information.
}
\description{
Weight of Evidence and Information Value for a categorical variable
}
\examples{
[TODO]
}
