% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{binTableContinuous}
\alias{binTableContinuous}
\title{Weight of Evidence and Information Value for a continuous variable}
\usage{
binTableContinuous(df, x, y, p = 0.05, verbose = FALSE)
}
\arguments{
\item{df}{Dataframe containing the two columns of data}

\item{x}{Name of the continuous variable as a string}

\item{y}{Name of the boolean response as a string}

\item{p}{Parameter for the Conditional Recuresive Tree used (see \code{partykit})}

\item{verbose}{Boolean to add additional information. Default is \code{FALSE}}
}
\value{
A list with the total Information Value, and a tibble containing the bins with detailed
 information.
}
\description{
Weight of Evidence and Information Value for a continuous variable
}
\examples{
[TODO]
}
