% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.R
\name{categoriseFromWoE}
\alias{categoriseFromWoE}
\title{Create a table that categorises a variable into the bins described by a WoE table}
\usage{
categoriseFromWoE(df, varName, woeTable, verbose = FALSE)
}
\arguments{
\item{df}{Dataframe containing the at least one column named `varName`}

\item{varName}{Name of the variable (string format) which will be binned follwoing its Weight of Evidence}

\item{woeTable}{Tibble containing the Weight of Evidence table created by WoETable for `varName`}

\item{verbose}{Boolean to add additional information. Default is \code{FALSE}}
}
\value{
Tibble of categories containing 0/1
}
\description{
This function takes a WoE table of a particular variable describing its bins, and takes the observations of the
variable. It creates a tibble of variables (one per bin) containing 0 or 1 whether the original observations
fall into the relevant bin.
}
