% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{chileancredit}
\alias{chileancredit}
\title{Chilean Credit Data}
\format{Data frame with 7,702 rows and 19 columns.}
\description{
A simulated dataset based on six months of information collected by a Chilean Bank
whose objective was to develop a credit scoring model to determine the probability
of default within the next 12 months. The target 
variable is FlagGB, which represents the binary status of default (0) and not default(1).
}
\details{
\itemize{
  \item CustomerId. Customer Identifier.
  \item TOB. Time on books in months since first account was open.
  \item IncomeLevel. Income level from 0 (Low) to 5 (High).
  \item Bal. Outstanding balance.
  \item MaxDqBin. Max. delinquency bin. 0:No Dq., 1:1-29 ... 6:150-179.
  \item MtgBal. Mortgage outstanding balance at the Credit Bureau.
  \item NonBankTradesDq. Number of non-bank delinquent trades.
  \item FlagGB. 1: Good, 0: Bad.
  \item FlagSample. Training and testing sample indicator (1:75\%,0:25\%).
  }
}

