% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{ensureLogical}
\alias{ensureLogical}
\title{Ensure that the content of df[,y] is boolean}
\usage{
ensureLogical(df, y, verbose = FALSE)
}
\arguments{
\item{y}{}
}
\value{
Reformatted data frame
}
\description{
This function is not exported and used in WoE table creation
}
