% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_etc.R
\name{plot_div}
\alias{plot_div}
\title{Divergence plot}
\usage{
plot_div(truth, pred)
}
\arguments{
\item{truth}{factor like TRUE/FALSE}

\item{pred}{predicted value}
}
\value{
Divergence statistics
}
\description{
Divergence plot
}

