% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_etc.R
\name{plot_ks}
\alias{plot_ks}
\title{kolmogorov-smirnov plot}
\usage{
plot_ks(bad_score, n_bad_score, bad_txt = "BAD", n_bad_txt = "NORM")
}
\arguments{
\item{bad_score}{bad_score scores explain BAD}

\item{n_bad_score}{n_bad_score scores explain NORMAL}

\item{bad_txt}{bad_txt text for BAD}

\item{n_bad_txt}{n_bad_txt text for NORMAL}
}
\value{
maximum difference in cumulative fraction
}
\description{
kolmogorov-smirnov plot
}

