% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdl_etc.R
\name{psi_f}
\alias{psi_f}
\title{PSI(Population Stability Index)}
\usage{
psi_f(base_vec, eval_vec, base_ym, eval_ym)
}
\arguments{
\item{base_vec}{base_vec used for reference PSI}

\item{eval_vec}{eval_vec used for target PSI}

\item{base_ym}{base_ym reference ymd}

\item{eval_ym}{eval_ym target ymd}
}
\description{
PSI(Population Stability Index)
}

