% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{smbinning.gen2}
\alias{smbinning.gen2}
\title{Utility to generate a new characteristic}
\usage{
smbinning.gen2(df, ivout, extend = FALSE)
}
\arguments{
\item{df}{Dataset for new characteristic.}

\item{ivout}{An object generated after \code{smbinning}.}

\item{extend}{if TRUE, overflowed input value will be merged to highest or lowest bins.}
}
\value{
A data frame with the binned version of the characteristic analyzed with \code{smbinning}.
}
\description{
It generates a vector with a new predictive characteristic after the binning process.
}
\author{
Heewon Jeon
}

