% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{smbinning.sql}
\alias{smbinning.sql}
\title{SQL Code}
\usage{
smbinning.sql(ivout)
}
\arguments{
\item{ivout}{An object generated by \code{smbinning}.}
}
\value{
A text with the SQL code for binning.
}
\description{
It outputs a SQL code to facilitate the generation of new binned characetristic 
in a SQL environment. User must define table and new characteristic name.
}
\examples{
# Load library and its dataset
library(smbinning)

# Example 1: Binning a numeric variable
result=smbinning(df=chileancredit,y="fgood",x="cbs1") # Run and save result
smbinning.sql(result)

# Example 2: Binning for a factor variable
result=smbinning.factor(df=chileancredit,x="inc",y="fgood",maxcat=11)
smbinning.sql(result)

# Example 3: Customized binning for a factor variable
result=smbinning.factor.custom(
  df=chileancredit,x="inc",y="fgood",
  c("'W01','W02'","'W03','W04','W05'",
    "'W06','W07'","'W08','W09','W10'"))
smbinning.sql(result)
}
