% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{smbsimdf3}
\alias{smbsimdf3}
\title{Monotonic Binning Sample Data}
\format{Data frame with 1,000 rows and 4 columns.}
\description{
A simulated dataset used to illustrate the application of model ranking.
}
\details{
\itemize{
  \item fgood1: Default (0), Not Default (1) for Numeric Variable 1.
  \item chr1: Numeric variable 1.
  \item chr2: Numeric variable 2.
  \item chr3: Numeric variable 3.
  }
}
