% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{binTableCategorical}
\alias{binTableCategorical}
\title{Weight of Evidence and Information Value for a categorical variable}
\usage{
binTableCategorical(df, x, y, maxCategories = 5, verbose = FALSE)
}
\arguments{
\item{df}{Dataframe containing the two columns of data}

\item{x}{Name of the categorical variable as a string}

\item{y}{Name of the logical response variable as a string.}

\item{maxCategories}{maximum number of accepted categories. Default is 5.}

\item{verbose}{Boolean to add additional information. Default is \code{FALSE}}
}
\value{
A list with the total Information Value, and a tibble containing the bins with detailed
 information.
}
\description{
This function calculates the Weight of Evidence and Infomation value for a particular response.
The variable tested is categorical (factors). The response is assumed to be logical. If not, if
the response takes only 2 values (either factor or numerical), it will be transformed into
logical values. `verbose = TRUE` shows the conversion.
}
\examples{
[TODO]
}
