% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{BinTableContinuous}
\alias{BinTableContinuous}
\title{Weight of Evidence and Information Value for a continuous variable}
\usage{
BinTableContinuous(df, x, y, p = 0.05, verbose = FALSE)
}
\arguments{
\item{df}{Dataframe containing the two columns of data}

\item{x}{Name of the continuous variable as a string}

\item{y}{Name of the boolean response as a string}

\item{p}{Parameter for the Conditional Recuresive Tree used (see `partykit`)}

\item{verbose}{Boolean to add additional information. Default is FALSE}
}
\value{
A tibble containing the bins with detailed information. The total information value
is obtained by summing the `IV` column (sum(resultTibble$IV))
}
\description{
Weight of Evidence and Information Value for a continuous variable
}
\examples{
[TODO]
}
