% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{chileancredit}
\alias{chileancredit}
\title{Chilean Credit Data}
\format{Data frame with 18,718 rows and 28 columns.}
\description{
A loan dataset where the target variable is FlagGB, 
which represents the binary status of default (0) and not default (1) after 12 months.
}
\details{
\itemize{
  \item LnCntRel: Number of owners.
  \item LnTOB: Time on books (months) since open.
  \item LnAutopay: Type of payment.
  \item CuScore1: Credit score 1.
  \item CuScore2: Credit score 2.
  \item CuScore3: Credit score 3.
  \item CuDpBal: Current amount of deposits.
  \item CuFlagSvCD: Number of savings and CDs.
  \item CuLnUnCnt: Number of unsecured loans.
  \item CuLnScCnt: Number of secured loans.
  \item CuMgCnt: Number of Mortgages.
  \item CuTOB: Time on books (months) since first account.
  \item CuDpTOB: Time on books (months) since first deposit account.
  \item CuDpPct12M: Growth of deposits (Percentage) in the last 12 months.
  \item CuDpDiff12M: Growth of deposits (Amount) in the last 12 months.
  \item CuPOSCntAvg12M: Average number of purchases in the last 12 months.
  \item CuPOSAmtAvg12M: Average amount of purchases in the last 12 months.
  \item CuPOSFq12M: Frequency of purchases in the last 12 months.
  \item CuDpAvgBal12M: Average amount of deposits in the last 12 months.
  \item CuDDCntAvg12M: Average number of electronic deposits in the last 12 months.
  \item CuDDAmtAvg12M: Average amount of electronic in the last 12 months.
  \item CuDDFq12M: Frequency of electronic deposits in the last 12 months.
  \item CuWealth: Income level.
  \item CuODRecency: Most recent overdraft.
  \item CuODCnt12M: Average number of overdrafts in the last 12 months.
  \item CuODFq12M: Frequency of overdrafts in the last 12 months.
  \item FlagGB: Default indicator where 0 means default.
  \item Rnd: Random number to select testing and training samples.
  }
}
