% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{chileancredit}
\alias{chileancredit}
\title{Chilean Credit Data}
\format{Data frame with 10,000 rows and 23 columns.}
\description{
A simulated dataset where the target variable is fgood, 
which represents the binary status of default (0) and not default (1).
}
\details{
\itemize{
  \item fgood: Default (0), Not Default (1).
  \item cbs1: Credit score 1.
  \item cbs2: Credit score 2.
  \item cbs3: Credit score 3.
  \item cbinq: Number of inquiries.
  \item cbline: Number of credit lines.
  \item cbterm: Number of term loans.
  \item cblineut: Line utilization (0-100).
  \item cbtob: Number of years on file.
  \item cbdpd: Indicator of days past due on bureau (Yes, No).
  \item cbnew: Number of new loans.
  \item pmt: Type of payment (M: Manual, A: Autopay, P: Payroll).
  \item tob: Time on books (Years).
  \item dpd: Level of delinquency (No, Low, High).
  \item dep: Amount of deposits own by customer.
  \item dc: Number of debit card transactions.
  \item od: Number of overdrafts.
  \item home: Home ownership indicator (Yes, No).
  \item inc: Level of income.
  \item dd: Number of direct deposits per month.
  \item online: Indicator of active online (Yes, No).
  \item rnd: Random number to select testing and training samples.
  \item period: Factor that indicates the year/month of the data.
  }
}
