% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{smbinning.scoring.gen}
\alias{smbinning.scoring.gen}
\title{Generation of Score and Its Weights}
\usage{
smbinning.scoring.gen(smbscaled, dataset)
}
\arguments{
\item{smbscaled}{Object generated using \code{smbinning.scaling}.}

\item{dataset}{A data frame.}
}
\value{
The command \code{smbinning.scoring} generates a data frame with the final scaled Score and its
corresponding scaled weights per characteristic.
}
\description{
After applying \code{smbinning.scaling} to the model, \code{smbinning.scoring} generates a data frame 
with the final Score and additional fields with the points assigned to each characteristic so the user
can see how the final score is calculated. Example shown on \code{smbinning.scaling} section.
}
