% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smbinning.R
\name{smbinning.scoring.sql}
\alias{smbinning.scoring.sql}
\title{Generation of SQL Code After Scaled Model}
\usage{
smbinning.scoring.sql(smbscaled)
}
\arguments{
\item{smbscaled}{Object generated using \code{smbinning.scaling}.}
}
\value{
The command \code{smbinning.scoring.sql} generates a SQL code to implement the model the model in SQL.
}
\description{
After applying \code{smbinning.scaling} to the model, \code{smbinning.scoring.sql} generates a SQL code 
that creates and updates all variables present in the scaled model. Example shown on \code{smbinning.scaling} section.
}
